upload to aws in new class
